multiple reasons + avoid duplicate code
